#10<Feat>: MainView view Update
